refactor: split Windows clipboard helpers
